refactor: merge when-false-explicit fixture into when-event
Consolidate the when-false-explicit test scenario into the when-event
fixture as both test f-when hydration behavior. Adds a test-element-false
element with an explicit someprop=false to the existing fixture.
