+++ packages/fast-html/test/fixtures/when-event/main.ts
@@ -1,5 +1,5 @@
 import { RenderableFASTElement, TemplateElement } from "@microsoft/fast-html";
-import { FASTElement } from "@microsoft/fast-element";
+import { FASTElement, observable } from "@microsoft/fast-element";
 
 // This element intentionally does NOT define "serverOnly" as a property.
 // The f-when condition references "serverOnly" which only exists in server state,
@@ -14,6 +14,18 @@ class TestElement extends FASTElement {
 }
 RenderableFASTElement(TestElement).defineAsync({
     name: "test-element",
+    templateOptions: "defer-and-hydrate",
+});
+
+// This element explicitly defines "someprop" as false.
+// The server rendered with someprop=true (from JSON), so SSR content is present.
+// During hydration, the client value is false, which should be respected.
+class TestElementFalse extends FASTElement {
+    @observable
+    someprop: boolean = false;
+}
+RenderableFASTElement(TestElementFalse).defineAsync({
+    name: "test-element-false",
     templateOptions: "defer-and-hydrate",
 });
