--- packages/fast-html/test/fixtures/when-false-explicit/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US">
-    <head>
-        <meta charset="utf-8" />
-        <title></title>
-        <script type="module" src="./main.ts"></script>
-    </head>
-    <body>
-        <f-template name="test-element">
-            <template><f-when value="{{someprop}}">anything, really</f-when></template>
-        </f-template>
-        <!-- SSR: someprop was true on server, content is rendered.
-             Client element has someprop=false, so content should be removed after hydration. -->
-        <test-element id="false-prop">
-            <template shadowrootmode="open"><!--fe-b$$start$$0$$MRl5Rw6tl3$$fe-b-->anything, really<!--fe-b$$end$$0$$MRl5Rw6tl3$$fe-b--></template>
-        </test-element>
-    </body>
-</html>
