--- packages/fast-html/test/fixtures/when-false-explicit/when-false-explicit.html
@@ -1,1 +1,0 @@
-<f-when value="{{someprop}}">anything, really</f-when>
