--- packages/fast-html/test/fixtures/when-false-explicit/when-false-explicit.json
@@ -1,3 +1,0 @@
-{
-    "someprop": true
-}
